refactor: moving route components into features
