+++ dqs-app/app/onboarding/page3.tsx
@@ -3,7 +3,7 @@ import {TouchableOpacity, Text} from 'react-native';
 import {router} from 'expo-router';
 import {TwContainer} from "@/core/components/TwContainer";
 import {TwText} from "@/core/components/TwText";
-import {Score} from "@/app/(app)/scores/components/Score";
+import {Score} from "@/features/scores/components/Score";
 import {FoodCat} from "@/core/enums";
 import {maxScores} from "@/core/constants";
 import {useState} from "react";
