+++ dqs-app/features/progress/components/DaySummary.tsx
