+++ dqs-app/features/scores/components/Score.tsx
@@ -3,8 +3,8 @@ import { Pressable } from "react-native";
 import { FoodCat } from "@/core/enums";
 import type { PossibleSingleServingScores } from "@/core/types";
 import React from "react";
-import {ScoreServing} from "@/app/(app)/scores/components/ScoreServing";
-import {ScoreLabel} from "@/app/(app)/scores/components/ScoreLabel";
+import {ScoreServing} from "@/features/scores/components/ScoreServing";
+import {ScoreLabel} from "@/features/scores/components/ScoreLabel";
 
 type ScoreProps = {
     text: string;
