+++ dqs-app/features/scores/components/ScoreLabel.tsx
